+++ exts/omni.isaac.lab_quadruped_tasks/omni/isaac/lab_quadruped_tasks/unitree_go2/__init__.py
@@ -5,17 +5,56 @@ Copyright (c) 2024, Felipe Mohr Santos
 
 import gymnasium as gym
 
-from omni.isaac.lab_quadruped_tasks import quadruped_env_cfg
+from . import agents, go2_env_cfg
 
-from . import agents
+##
+# Register Gym environments.
+##
 
-# Register Gym environments
+############################
+# Go2 Blind Flat Environment
+############################
+
 gym.register(
-    id="Isaac-Quadruped-Go2-Flat-v0",
+    id="Isaac-Quadruped-Go2-Blind-Flat-v0",
     entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": quadruped_env_cfg.QuadrupedEnvCfg,
-        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.QuadrupedGo2PPORunnerCfg,
+        "env_cfg_entry_point": go2_env_cfg.Go2BlindFlatEnvCfg,
+        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.Go2BlindFlatPPORunnerCfg,
+    },
+)
+
+gym.register(
+    id="Isaac-Quadruped-Go2-Blind-Flat-Play-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": go2_env_cfg.Go2BlindFlatEnvCfg_PLAY,
+        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.Go2BlindFlatPPORunnerCfg,
+    },
+)
+
+#############################
+# Go2 Blind Rough Environment
+#############################
+
+gym.register(
+    id="Isaac-Quadruped-Go2-Blind-Rough-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": go2_env_cfg.Go2BlindRoughEnvCfg,
+        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.Go2BlindRoughPPORunnerCfg,
+    },
+)
+
+gym.register(
+    id="Isaac-Quadruped-Go2-Blind-Rough-Play-v0",
+    entry_point="omni.isaac.lab.envs:ManagerBasedRLEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": go2_env_cfg.Go2BlindRoughEnvCfg_PLAY,
+        "rsl_rl_cfg_entry_point": agents.rsl_rl_cfg.Go2BlindRoughPPORunnerCfg,
     },
 )